Fix crash due to bad handling of self as delegate/datasource
This crash occurs when quickly double-clicking an item or running the unit tests on Sierra. The reference to self is not valid in animation handler, since it has been dealloced. Instead of overriding the setter and getter, go through the super class (NSOutlineView).

Also update the unit tests to work and pass with new changes.